MDEV-13915: STOP SLAVE takes very long time on a busy system
The problem is that a parallel replica would not immediately stop
running/queued transactions when issued STOP SLAVE. This would lead
to long periods to wait for all waiting transactions to finish.

This patch updates a parallel replica to try and abort immediately
and roll-back any ongoing transactions. There are two exceptions:
   1. The transaction is next-to-commit and it has already modified
      non-transactional table and cannot be safely rolled back.
   2. We are at the commit step of the transaction.

Reviewed By:
============
<TODO>